Removing duplicate test code: duplicate_particles done.
