+++ testing/check_presence_duplicate_particles.py
@@ -1,84 +1,40 @@
 #!/bin/env python
-from tqdm import tqdm
+
+# Retrieve helper functions, without having to define an __init__.py 
+import sys
+sys.path.append('../toolbox')
+from helper_functions import read_particles
+
 from mpi4py import MPI
 comm = MPI.COMM_WORLD
 comm_rank = comm.Get_rank()
 comm_size = comm.Get_size()
 
-import h5py
 import numpy as np
-
-import virgo.mpi.util
 import virgo.mpi.parallel_hdf5 as phdf5
 import virgo.mpi.parallel_sort as psort
 import virgo.mpi.gather_array as gather_array
 
-def read_particles(filenames, nr_local_subhalos):
+def check_duplicate_particles(basedir, hbt_nr):
     """
-    Read in the particle IDs belonging to the subhalos on this MPI
-    rank from the specified SubSnap files. Returns a single array with
-    the concatenated IDs from all local subhalos in the order they
-    appear in the SubSnap files.
-    """
+    This function checks for the presence of particles that are bound to more
+    than one HBT subgroup. 
 
-    # First determine how many subhalos are in each SubSnap file
-    if comm_rank == 0:
-        subhalos_per_file = []
-        nr_files = 1
-        file_nr = 0
-        while file_nr < nr_files:
-            with h5py.File(filenames.format(file_nr=file_nr), "r") as infile:
-                subhalos_per_file.append(infile["Subhalos"].shape[0])
-                nr_files = int(infile["NumberOfFiles"][...])
-            file_nr += 1
-    else:
-        subhalos_per_file = None
-        nr_files = None
-    nr_files = comm.bcast(nr_files)
-    subhalos_per_file = np.asarray(comm.bcast(subhalos_per_file), dtype=int)
-    first_subhalo_in_file = np.cumsum(subhalos_per_file) - subhalos_per_file
-    
-    # Determine offset to first subhalo this rank reads
-    first_local_subhalo = comm.scan(nr_local_subhalos) - nr_local_subhalos
+    Parameters
+    ----------
+    basedir: str    
+        Location of the HBT catalogues.
+    hbt_nr : int
+        Snapshot index to test.
 
-    # Loop over all files
-    particle_ids = []
-    for file_nr in range(nr_files):
-
-        # Find range of subhalos this rank read from this file
-        i1 = first_local_subhalo - first_subhalo_in_file[file_nr]
-        i2 = i1 + nr_local_subhalos
-        i1 = max(0, i1)
-        i2 = min(subhalos_per_file[file_nr], i2)
-
-        # Read subhalo particle IDs, if there are any in this file for this rank
-        if i2 > i1:
-            with h5py.File(filenames.format(file_nr=file_nr), "r") as infile:
-                particle_ids.append(infile["SubhaloParticles"][i1:i2])
-
-    if len(particle_ids) > 0:
-        # Combine arrays from different files
-        particle_ids = np.concatenate(particle_ids)
-        # Combine arrays from different subhalos
-        particle_ids = np.concatenate(particle_ids)        
-    else:
-        # Some ranks may have read zero files
-        particle_ids = None
-    particle_ids = virgo.mpi.util.replace_none_with_zero_size(particle_ids, comm=comm)
-    
-    return particle_ids
-
-def sort_hbt_output(basedir, hbt_nr):
-
-    """
-    This reorganizes a set of HBT SubSnap files into a single file which
-    contains one HDF5 dataset for each subhalo property. Subhalos are written
-    in order of TrackId.
-
-    Particle IDs in groups can be optionally copied to the output.
+    Returns
+    -------
+    total_number_duplicates: int
+        Number of particles that are shared bound to more than one subgroup.
+    number_unique_subhalos_with_duplicate_particles : int
+        Number of unique subgroups that share particles.
     """
 
-    # Read in the input subhalos
     if comm_rank == 0:
         print(f"Testing presence of duplicate particles in HBTplus at snapshot index {hbt_nr}.")
 
@@ -150,12 +106,21 @@ def sort_hbt_output(basedir, hbt_nr):
     if comm_rank == 0:
         print(f"{total_number_duplicates} particles out of {total_number_particles} are duplicate.") 
     
-    # Retrieve the tracks that share particles, if any
+    # Retrieve the subhalos that share particles, if any
     if total_number_duplicates != 0:
-        tracks_with_shared_particles = gather_array.allgather_array(particle_trackids[next_particle_ids == particle_ids], comm=comm)
+        # Number of unique subhalos that share particles
+        subhalos_with_shared_particles = gather_array.allgather_array(particle_trackids[next_particle_ids == particle_ids], comm=comm)
+        number_unique_subhalos_with_duplicate_particles = len(np.unique(subhalos_with_shared_particles))
+
+        # Number of subhalos we tested 
         global_number_subhalos = comm.allreduce(local_nr_subhalos)
+        
         if comm_rank == 0:
-           print(f"{len(np.unique(tracks_with_shared_particles))} unique Tracks out of {global_number_subhalos} share particles.")
+           print(f"{number_unique_subhalos_with_duplicate_particles} unique subhalos out of {global_number_subhalos} share particles.")
+    else:
+        number_unique_subhalos_with_duplicate_particles = 0
+
+    return total_number_duplicates, number_unique_subhalos_with_duplicate_particles
 
 if __name__ == "__main__":
 
@@ -167,4 +132,4 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
 
-    sort_hbt_output(**vars(args))
+    check_duplicate_particles(**vars(args))
